Refactor campaign form into hook and section components
