+++ packages/ui/src/components/cms/marketing/campaign/CampaignFormToast.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+import { Toast } from "../../../atoms";
+import type { CampaignFormToastState } from "./useCampaignForm";
+
+interface CampaignFormToastProps extends CampaignFormToastState {
+  onClose: () => void;
+}
+
+export function CampaignFormToast({ open, message, onClose }: CampaignFormToastProps) {
+  return <Toast open={open} message={message} onClose={onClose} />;
+}
+
+export default CampaignFormToast;
